+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * [Improved] Handle exceptions inside `postProcess`
   * [Improved] #568 ext-dnd: Auto-expanding of collapsed nodes should also work 
     when dropping is not allowed
+  * [Improved] #567 ext-dnd: fix revert position
   * [Fixed] #562 Node span tag leaks outside table cell
   * [Fixed] #563 `tree.reactivate(false)` sets fancytree-treefocus and `tree.reactivate(true)` 
     doesn't set keyboard focus